% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boe-package.R
\docType{package}
\name{boe-package}
\alias{boe}
\alias{boe-package}
\title{boe: Download Data from the 'Bank of England' Statistical Database}
\description{
Provides functions to download and tidy statistical data published by the 'Bank of England' \url{https://www.bankofengland.co.uk}. Covers Bank Rate, 'SONIA', gilt yields, exchange rates, mortgage rates, mortgage approvals, consumer credit, and money supply. Series are fetched from the 'Bank of England Interactive Statistical Database' using its CSV endpoint. Data is cached locally between sessions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://charlescoverdale.github.io/boe/}
  \item \url{https://github.com/charlescoverdale/boe}
  \item Report bugs at \url{https://github.com/charlescoverdale/boe/issues}
}

}
\author{
\strong{Maintainer}: Charles Coverdale \email{charlesfcoverdale@gmail.com}

}
\keyword{internal}
