% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_xml}
\alias{get_xml}
\title{Retrieve the XML content}
\usage{
get_xml(query)
}
\arguments{
\item{query}{A query to BOE}
}
\description{
Look up on the internet and get the content
}
\examples{
id <- sumario_xml(format(as.Date("2017/10/02", "\%Y/\%m/\%d"), "\%Y\%m\%d"))
get_xml(query_xml(id))
}
\seealso{
query_xml
}
