% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_date_boe.R
\docType{data}
\name{last_date_boe}
\alias{last_date_boe}
\alias{last_boe}
\title{Last day of BOE}
\format{
An object of class \code{Date} of length 1.
}
\usage{
last_date_boe()

last_boe
}
\value{
A date in character format \%Y-\%m-\%d.
}
\description{
Returns the date of the last BOE under normal circumstances.
}
\examples{
last_boe()
}
\keyword{datasets}
