% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leg_cons.R
\name{lex}
\alias{lex}
\title{Search consolidated legislature.}
\usage{
lex(query, from, to = Sys.Date(), offset = NULL, limit = -1)
}
\arguments{
\item{query}{Search parameters.}

\item{from}{Earliest date from which the legislation is collected.}

\item{to}{Latest date from which the legislation is collected.}

\item{offset}{Return from this results.}

\item{limit}{How many results return. The default returns all of them.}
}
\description{
Search consolidated legislature.
}
\examples{
lex("inteligencia artificial", Sys.Date() - 365)
}
