% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leg_cons.R
\name{lex_id}
\alias{lex_id}
\title{Download by CVE}
\usage{
lex_id(id, metadatos = FALSE, analisis = FALSE, ELI = FALSE)
}
\arguments{
\item{id}{CVE of the document.}

\item{metadatos}{Logical value if metadata should be retrieved.}

\item{analisis}{Logical value if analysis should be retrieved.}

\item{ELI}{Logical value if ELI metadata should be retrieved.}
}
\value{
Something
}
\description{
Download by CVE
}
\examples{
lex_id("BOE-A-2000-17579")
}
