% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{open_publications}
\alias{open_publications}
\title{Open publications on the browser}
\usage{
open_publications(sumario)
}
\arguments{
\item{sumario}{A tidy sumario.}
}
\value{
A tab should open on your default browser with the pdf of the publications.
}
\description{
This function opens urls in your browser.
}
\examples{
\donttest{
sumario <- retrieve_sumario(Sys.Date())
url_publications(sumario[1:10, ])}
}
