% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_htm}
\alias{query_htm}
\title{Build a query for the webpage}
\usage{
query_htm(cve)
}
\arguments{
\item{cve}{The CVE of the document you want.}
}
\value{
A query url.
}
\description{
Build a query for the webpage
}
\examples{
cve <- sumario_cve("2017", "117")
query_htm(cve)
}
