% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_document}
\alias{retrieve_document}
\title{Retrieve information of a publication}
\usage{
retrieve_document(id)
}
\arguments{
\item{id}{Character with the code of the document.}
}
\description{
Tidy data from any document published on the BOE and BORME from the
\href{https://boe.es}{BOE}.
}
\examples{
xml  <- get_xml(query_xml("BOE-B-2017-5"))
xml2 <- get_xml(query_xml("BOE-A-2017-5"))
xml3 <- get_xml(query_xml("BOE-S-2017-5"))
xml4 <- get_xml(query_xml("BORME-S-2020-108"))
id <- "BOE-A-2020-12109"
df <- retrieve_document(id)
}
\seealso{
Other functions to retrieve documents: 
\code{\link{retrieve_sumario}()}
}
\concept{functions to retrieve documents}
