% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_sumario}
\alias{retrieve_sumario}
\title{Retrieve the sumario}
\usage{
retrieve_sumario(date)
}
\arguments{
\item{date}{A Date of the \emph{sumario}.}
}
\value{
A data.frame with one line for each publication.
}
\description{
Obtain a sumario and tidy it in a table.
}
\examples{
\donttest{retrieve_sumario(Sys.Date())}
}
\seealso{
tidy_sumario to learn about the format of the output
}
