% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{sumario_cve}
\alias{sumario_cve}
\title{Create the number of the \emph{sumario}}
\usage{
sumario_cve(year, number, journal = "BOE")
}
\arguments{
\item{year}{Character or numeric value of the year of the summary in YYYY format.}

\item{number}{Number of the summary in NNN format.}

\item{journal}{Either BOE or BORME.}
}
\value{
A character vector with the CVE of the \emph{sumario}.
}
\description{
Create the number of the \emph{sumario}
}
\examples{
sumario_cve(2019, 242)
}
\seealso{
\code{\link[=sumario_nbo]{sumario_nbo()}} if you want to retrieve the \emph{sumario} by date and don't know the CVE.

Other code generators: 
\code{\link{element}},
\code{\link{sumario_nbo}()}
}
\concept{code generators}
