% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{sumario_nbo}
\alias{sumario_nbo}
\alias{sumario_xml}
\title{Create the number for \emph{sumario}}
\usage{
sumario_nbo(date, journal = "BOE")

sumario_xml(date)
}
\arguments{
\item{date}{Date of the \emph{sumario}}

\item{journal}{Either BOE or BORME.}
}
\value{
A character vector
}
\description{
Creates the id for a \emph{sumario} by the date it was published.
}
\section{Functions}{
\itemize{
\item \code{sumario_xml()}: For compatibility with previous version

}}
\examples{
sumario_nbo(Sys.Date())
sumario_nbo(format(as.Date("2009/01/01", "\%Y/\%m/\%d"), "\%Y\%m\%d"), journal = "BORME")
}
\seealso{
The id is different from the CVE, which can be created with \code{\link[=sumario_cve]{sumario_cve()}}.

Other code generators: 
\code{\link{element}},
\code{\link{sumario_cve}()}
}
\concept{code generators}
