% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{sumario_suplementos}
\alias{sumario_suplementos}
\title{Supplementary summaries}
\usage{
sumario_suplementos(year, number, type = "N")
}
\arguments{
\item{year}{Character or numeric value of the year of the summary in YYYY format.}

\item{number}{Number of the summary in NNN format.}

\item{type}{Either J or N. J for \strong{judicial} or N for \strong{notificaciones}}
}
\value{
A CVE of the document
}
\description{
Creates the CVE of a summary of the supplements, either the judicial or notifications.
These are only available for 3 months.
}
\examples{
sumario_suplementos(2023, 1)
}
