% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{sumario_xml}
\alias{sumario_xml}
\title{Create the id for \emph{sumario}}
\usage{
sumario_xml(date, journal = "BOE")
}
\arguments{
\item{date}{Date of the \emph{sumario}}

\item{journal}{Either BOE or BORME.}
}
\value{
A character vector
}
\description{
Create the id for \emph{sumario}
}
\examples{
sumario_xml(Sys.Date())
sumario_xml(format(as.Date("2009/01/01", "\%Y/\%m/\%d"), "\%Y\%m\%d"), journal = "BORME")
}
\seealso{
\url{sumario_nbo}

Other code generator: 
\code{\link{element}},
\code{\link{sumario_nbo}()}
}
\concept{code generator}
