% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{suplemento_cve}
\alias{suplemento_cve}
\title{Supplement CVE}
\usage{
suplemento_cve(year = 2023, number, type = "J")
}
\arguments{
\item{year}{Character or numeric value of the year of the summary in YYYY format.}

\item{number}{Number of the summary in NNN format.}

\item{type}{Either J or N. J for \strong{judicial} or N for \strong{notificaciones}}
}
\value{
A CVE.
}
\description{
Creates the CVE of a supplement, either the judicial or notifications.
These are only available for 3 months.
}
\examples{
suplemento_cve(number = 1)
}
