% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_sumario}
\alias{tidy_sumario}
\title{Tidy sumario xml}
\usage{
tidy_sumario(x)
}
\arguments{
\item{x}{A XML file.}
}
\value{
A data.frame with one line for each publication.
Including date, \emph{sumario} number, section, section text,
departament, departament number, epigraph, brief text,
publication code and number of pages of the pdf.
}
\description{
Cleans in a tidy format the sumario file
}
\examples{
id <- sumario_xml(format(as.Date("2017/10/02", "\%Y/\%m/\%d"), "\%Y\%m\%d"))
sumario_file <- get_xml(query_xml(id))
m <- tidy_sumario(sumario_file)
head(m)
tail(m)
}
