% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{url_publications}
\alias{url_publications}
\title{Url to the publications}
\usage{
url_publications(sumario)
}
\arguments{
\item{sumario}{A tidy sumario.}
}
\value{
A character vector with the urls to get the publications in pdf format.
}
\description{
Transform the date and publication code to the urls to the publications.
}
\examples{
\donttest{
sumario <- retrieve_sumario(Sys.Date())
url_publications(sumario[1:10, ])
}
}
\seealso{
\url{retrieve_sumario} to obtain \emph{sumarios}, and \url{tidy_sumario}
to know the expected input.
}
