% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOE-package.R
\docType{package}
\name{BOE-package}
\alias{BOE}
\alias{BOE-package}
\title{BOE: Get the Boletin Oficial del Estado of Spain.}
\description{
Queries the online site of the BOE for data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/llrs/BOE}
  \item Report bugs at \url{https://github.com/llrs/BOE/issues}
}

}
\author{
\strong{Maintainer}: Lluís Revilla Sancho \email{lluis.revilla@gmail.com} (\href{https://orcid.org/0000-0001-9747-2570}{ORCID})

}
\keyword{internal}
