% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{check_code}
\alias{check_code}
\title{Check code of documents}
\usage{
check_code(id)
}
\value{
A logical value.
}
\description{
Given an id check if it is valid.
}
\examples{
check_code("BOE-S-2014-242"))
# Will fail:
# check_code("BOE-S-2014")
# check_code("BOE-S-20141221")
}
