% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{check_code}
\alias{check_code}
\title{Check code of documents}
\usage{
check_code(id)
}
\arguments{
\item{id}{ID of the document (character).}
}
\value{
A logical value.
}
\description{
Given an id check if it is valid.
}
\examples{
check_code("BOE-S-20141006") # Normal way
check_code("BOE-S-2014-242") # Also accepted but not documented
# Will fail:
# check_code("BOE-S-2014")
}
