% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{element}
\alias{element}
\alias{disposicion}
\alias{anuncio}
\title{Elements: Disposición and anuncio}
\usage{
disposicion(year, number)

anuncio(year, number)
}
\arguments{
\item{year}{Character or numeric value of the year of the element in YYYY format.}

\item{number}{Character or numeric value of the element.}
}
\value{
A character vector
}
\description{
Elements: Disposición and anuncio
}
\section{Functions}{
\itemize{
\item \code{disposicion}: Create the number of the diposicion.

\item \code{anuncio}: Create the number of the anuncio.
}}

\examples{
disposicion(2019, 242)
anuncio(2019, 242)
}
