% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_date_boe.R
\name{last_date_boe}
\alias{last_date_boe}
\title{Last day of BOE}
\usage{
last_date_boe()
}
\value{
A date.
}
\description{
Returns the date of the last BOE under normal circumstances.
}
\examples{
last_date_boe()
}
