% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_htm}
\alias{query_htm}
\title{Build a query for the webpage}
\usage{
query_htm(id, journal = "BOE")
}
\arguments{
\item{id}{The id of the xml document you want.}

\item{journal}{Either BOE or BORME.}
}
\value{
A query url.
}
\description{
Build a query for the webpage
}
\examples{
id <- sumario_nbo("2017", "117")
query_htm(id)
}
