% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_htm}
\alias{query_htm}
\title{Build a query for the webpage}
\usage{
query_htm(id)
}
\arguments{
\item{id}{The id of the xml document you want.}
}
\description{
Build a query for the webpage
}
\examples{
id <- sumario_nbo("2017", "11117")
query_htm(id)
}
