% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_pdf}
\alias{query_pdf}
\title{Query a pdf from the BOE}
\usage{
query_pdf(year, month, day, code)
}
\arguments{
\item{year}{Character of the number of the year: YYYY}

\item{month}{Character of the number of the month: MM.}

\item{day}{Character of the number of the day: DD.}

\item{code}{Code of the publication to query.}
}
\description{
To query a pdf you must know the number of the piece you are looking for and
the date it was published.
}
\examples{
code <- sumario_nbo("2017", "237")
query_pdf("2017", "10", "02", code)
}
