% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_xml}
\alias{query_xml}
\title{Build a query for an XML}
\usage{
query_xml(id)
}
\arguments{
\item{id}{The id of the xml document you want.}
}
\value{
A query for the xml.
}
\description{
Build a query for an XML
}
\examples{
id <- sumario_nbo(format(as.Date("2017/10/02", "\%Y/\%m/\%d"), "\%Y\%m\%d"))
cve <- sumario_cve("2017", "237") # Same document but by the CVE
query_xml(id)
query_xml(cve)
}
\seealso{
\code{\link[=sumario_nbo]{sumario_nbo()}}
}
