% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_document}
\alias{retrieve_document}
\title{Retrieve information of a publication}
\usage{
retrieve_document(cve)
}
\arguments{
\item{cve}{Character with the code of the document.}
}
\description{
Tidy data from any document published on the BOE and BORME from the
\href{https://boe.es}{BOE}.
}
\examples{
cve <- "BOE-A-2020-12109"
df <- retrieve_document(cve)
}
\seealso{
Other functions to retrieve documents: 
\code{\link{retrieve_sumario}()}
}
\concept{functions to retrieve documents}
