% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{sumario_nbo}
\alias{sumario_nbo}
\title{Create the number for \emph{sumario}}
\usage{
sumario_nbo(date, journal = "BOE")
}
\arguments{
\item{date}{Date of the \emph{sumario}}

\item{journal}{Either BOE or BORME.}
}
\value{
A character vector
}
\description{
Create the number for \emph{sumario}
}
\examples{
sumario_nbo(Sys.Date())
sumario_nbo(format(as.Date("2009/01/01", "\%Y/\%m/\%d"), "\%Y\%m\%d"), journal = "BORME")
}
\seealso{
\code{\link[=sumario_cve]{sumario_cve()}}

Other code generator: 
\code{\link{element}},
\code{\link{sumario_cve}()}
}
\concept{code generator}
