% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{sumario_nbo}
\alias{sumario_nbo}
\title{Create the number of the \emph{sumario}}
\usage{
sumario_nbo(year, number, journal = "BOE")
}
\arguments{
\item{year}{Character or numeric value of the year of the summary in YYYY format.}

\item{number}{Number of the summary in NNN format.}

\item{journal}{Either BOE or BORME.}
}
\value{
A character vector
}
\description{
Create the number of the \emph{sumario}
}
\examples{
sumario_nbo(2019, 242)
}
\seealso{
Other code generator: 
\code{\link{element}},
\code{\link{sumario_xml}()}
}
\concept{code generator}
