% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{sumario_nbo}
\alias{sumario_nbo}
\title{Create the number of the sumario}
\usage{
sumario_nbo(year, number)
}
\arguments{
\item{year}{Character or numeric value of the year of the summary in YYYY format.}

\item{number}{Number of the summary in NNN format.}
}
\value{
A character vector
}
\description{
Create the number of the sumario
}
\examples{
sumario_nbo(2019, 242)
}
\seealso{
[sumario_xml]
}
