% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{sumario_xml}
\alias{sumario_xml}
\title{Create the id for sumario}
\usage{
sumario_xml(date)
}
\arguments{
\item{date}{Date of the sumario}
}
\value{
A character vector
}
\description{
Create the id for sumario
}
\examples{
sumario_xml(Sys.Date())
sumario_xml(format(as.Date("2009/01/01", "\%Y/\%m/\%d"), "\%Y\%m\%d"))
}
\seealso{
[sumario_nbo]
}
